Refactor ReservationService and related classes for improved token handling and order management; add OrderDTO and update payment service interfaces.

+++ src/main/java/ticketsystem/DomainLayer/order/ActiveOrder.java
@@ -6,13 +6,13 @@ import java.util.List;
 public class ActiveOrder {
 
     private int orderId;
-    private Integer userId;
+    private Long userId;
     private String sessionToken;
     private int eventId;
     private List<Ticket> tickets;
     private OrderStatus status;
 
-    public ActiveOrder(int orderId, int userId,String sessionToken, int eventId) {
+    public ActiveOrder(int orderId, Long userId,String sessionToken, int eventId) {
         this.orderId = orderId;
         this.userId = userId;
         this.sessionToken = sessionToken;
@@ -50,7 +50,7 @@ public class ActiveOrder {
         this.status = OrderStatus.COMPLETED;
     }
     
-    public int getUserId() {
+    public Long getUserId() {
         return this.userId;
     }
